+++ src/PhpSpreadsheet/Shared/CodePage.php
@@ -8,6 +8,7 @@ class CodePage
 {
     public const DEFAULT_CODE_PAGE = 'CP1252';
 
+    /** @var array<int, array<int, string>|string> */
     private static array $pageArray = [
         0 => 'CP1252', //    CodePage is not always correctly set when the xls file was saved by Apple's Numbers program
         367 => 'ASCII', //    ASCII